Move reshape cookies to common.
