+++ packages/common/src/index.ts
@@ -46,6 +46,7 @@ export { default as deriveBlockingStatus } from './utils/deriveBlockingStatus';
 export { default as getLegendDescription } from './utils/getLegendDescription';
 export { default as extractCookies } from './utils/extractCookies';
 export { default as extractReportData } from './utils/extractReportData';
+export { default as reshapeCookies } from './utils/reshapeCookies';
 export * from './worker/enums';
 export * from './utils/generateReports';
 export * from './cookies.types';
