+++ packages/report/src/dashboard/components/siteMapReport/layout.tsx
@@ -25,6 +25,7 @@ import {
   type CompleteJson,
   type LibraryData,
   noop,
+  reshapeCookies,
 } from '@google-psat/common';
 import {
   useSidebar,
@@ -39,7 +40,6 @@ import {
  */
 
 import SiteMapCookiesWithIssues from './sitemapCookiesWithIssues';
-import reshapeCookies from '../utils/reshapeCookies';
 import CookiesTab from './cookies';
 
 interface LayoutProps {
